+++ src/components/Loader/index.module.scss
