+++ src/components/Slider/index.js
