+++ src/components/Slideshow/Slide/index.js
