+++ src/pages/Account/AccountAddresses/index.js
@@ -1,6 +1,6 @@
 import { useAuthContext } from 'hooks/useAuthContext';
 
-import Button from 'common/Button';
+import Button from 'components/Button';
 
 import styles from './index.module.scss';
