+++ src/pages/Account/index.js
@@ -8,10 +8,10 @@ import AccountOrders from './AccountOrders';
 import AccountProfile from './AccountProfile';
 import AccountAddresses from './AccountAddresses';
 
-import Button from 'common/Button';
-import Loader from 'common/Loader';
-import Toast from 'common/Toast';
-import ToastMessage from 'common/ToastMessage';
+import Button from 'components/Button';
+import Loader from 'components/Loader';
+import Toast from 'components/Toast';
+import ToastMessage from 'components/ToastMessage';
 
 import styles from './index.module.scss';
