+++ src/pages/Addresses/AddAddress/index.js
@@ -3,9 +3,9 @@ import { useState, useEffect, useRef } from 'react';
 import { useAddress } from 'hooks/useAddress';
 import { useKeyDown } from 'hooks/useKeyDown';
 
-import Loader from 'common/Loader';
-import Toast from 'common/Toast';
-import ToastMessage from 'common/ToastMessage';
+import Loader from 'components/Loader';
+import Toast from 'components/Toast';
+import ToastMessage from 'components/ToastMessage';
 
 import styles from './index.module.scss';
