+++ src/pages/Cart/CartItem/index.js
@@ -2,7 +2,7 @@ import { Link, useParams } from 'react-router-dom';
 
 import { FaTrash, FaPlus, FaMinus } from 'react-icons/fa';
 
-import Card from 'common/Card';
+import Card from 'components/Card';
 
 import { formatNumber } from 'helpers/format';
 import { addItemPrice } from 'helpers/item';
