+++ src/pages/Collections/index.js
@@ -3,7 +3,7 @@ import { useParams, useNavigate } from 'react-router-dom';
 
 import { useCollection } from 'hooks/useCollection';
 
-import Loader from 'common/Loader';
+import Loader from 'components/Loader';
 
 import ProductCard from './ProductCard';
