+++ src/pages/Login/index.js
@@ -3,9 +3,9 @@ import { Link, useLocation } from 'react-router-dom';
 
 import { useLogin } from 'hooks/useLogin';
 
-import Loader from 'common/Loader';
-import Toast from 'common/Toast';
-import ToastMessage from 'common/ToastMessage';
+import Loader from 'components/Loader';
+import Toast from 'components/Toast';
+import ToastMessage from 'components/ToastMessage';
 
 import styles from './index.module.scss';
